Bugfix in helper.sh
Missing option `-f` in `init_broken_repo` for environments where `mv` is aliased to `mv -i`.
This bug is not easily detected in the jupyter notebook cell. It
requires interaction from the user (who has `mv -i` by default), but this is never seen in the cell output,
and it looks like the cell was executed correctly.